+++ h323plus/src/h323pdu.cxx
@@ -1539,7 +1539,7 @@ void H323SignalPDU::BuildQ931()
 
 void H323SignalPDU::PrintOn(ostream & strm) const
 {
-  int indent = strm.precision() + 2;
+  int indent = (int)strm.precision() + 2;
   strm << "{\n"
        << setw(indent+10) << "q931pdu = " << setprecision(indent) << q931pdu << '\n'
        << setw(indent+10) << "h225pdu = " << setprecision(indent);
